requirements.txt
pandas==1.2.5
numpy==1.20.3
wandb==0.13.5
scikit-learn==0.23.2
scipy==1.6.3
seaborn==0.11.1
setuptools==41.0.1
torchvision>=0.2.2
tqdm
wandb==0.13.5
